refactor: move EnvelopeProtoMapperImpl into common
