--- network/src/darwinMain/kotlin/com/wire/kalium/network/api/base/authenticated/message/OtrClientEntryMapper.kt
@@ -1,13 +1,0 @@
-package com.wire.kalium.network.api.base.authenticated.message
-
-import com.wire.kalium.protobuf.otr.ClientEntry
-import pbandk.ByteArr
-
-class OtrClientEntryMapper {
-
-    private val clientIdMapper = OtrClientIdMapper()
-    fun toOtrClientEntry(clientPayload: Map.Entry<String, ByteArray>): ClientEntry = ClientEntry(
-        client = clientIdMapper.toOtrClientId(clientPayload.key),
-        text = ByteArr(clientPayload.value),
-    )
-}
